requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
backoff==2.2.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
croniter==1.4.1
cryptography==38.0.4
databricks-sql-connector==2.0.3
dataclasses==0.6
distro==1.9.0
duckdb==0.3.4
duckdb-engine==0.1.11
exceptiongroup==1.2.2
fastapi==0.112.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gql==3.5.0
graphql-core==3.2.3
graphviz==0.18.2
greenlet==3.0.3
grpcio==1.66.0
grpcio-status==1.66.0
h11==0.14.0
halo==0.0.31
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.4
jiter==0.5.0
jsonschema==3.2.0
log-symbols==0.0.14
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
metricflow-lite==0.130.2
mo-dots==4.22.21108
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==4.22.21108
mo-logs==4.23.21108
monotonic==1.6
more-itertools==8.10.0
moz-sql-parser==4.40.21126
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.24.0
openai==0.28.0
opencv-python==4.10.0.84
oscrypto==1.3.0
packaging==24.1
pandas==1.4.4
pathspec==0.12.1
Pillow==8.4.0
pluggy==1.5.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.27.3
pyarrow==5.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycron==3.0.0
pycryptodomex==3.20.0
pydantic==1.10.18
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==22.1.0
pyrsistent==0.20.0
pytest==8.3.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
rudder-sdk-python==1.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==2.9.0
snowflake-sqlalchemy==1.2.3
spinners==0.0.24
SQLAlchemy==1.4.53
sqlalchemy-bigquery==1.11.0
sqlalchemy-redshift==0.8.1
sqlalchemy2-stubs==0.0.2a38
starlette==0.38.2
sympy==1.13.2
tabulate==0.8.9
termcolor==2.4.0
thrift==0.13.0
tomli==2.0.1
torch==1.10.0
torchvision==0.11.0
tqdm==4.66.5
transform==1.4.3
transform-tools==1.1.0
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
urllib3==1.26.14
uvicorn==0.30.6
uvloop==0.20.0
validators==0.33.0
watchfiles==0.23.0
websockets==12.0
xxhash==3.5.0
yacs==0.1.8
yamllint==1.35.1
yarl==1.9.4
